others code/old/hr_management_backend-main/requirements.txt
boto3==1.36.21
Django==5.1.6
django-extensions==3.2.3
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
django-environ==0.12.0
django-debug-toolbar==5.0.1